Fix for dark bands and white stripes on ZJ-58
The current USB/parallel-interface version of ZJ-58 available from AliExpress, presented in the system as /dev/usb/lp0 when connected via USB cable, prints dark bands and individual white lines when printing an image or a photo. It seems that either the original assumption that each individual empty line can be skipped using the faster "flush" command was wrong or that the firmware was updated in such a way that it is no longer possible to do that safely. Possibly the thermal printer is too fast for the rasterizer to keep up providing rasterized data?

The minimal lines to be either skipped or printed is now set to 24 as multiple lines were already being skipped in 24-line chunks where possible. It seems that packing non-empty lines to be printed into such 24-line chunks too is necessary to prevent a sort of black half-line artifacts. The new approach rasterizes some blank lines as a side effect but that normally noesn't slow down the printing process. Any trailing block of lines smaller than 24 lines is printed as it is, hoping that the printer won't screw up the end of the image knowing beforehand that no further data would follow.